--- ui/app/helpers/higher-order-components/with-token-tracker/index.js
@@ -1,1 +1,0 @@
-export { default } from './with-token-tracker.component'
