+++ ui/app/pages/confirm-approve/index.js
@@ -1,1 +1,1 @@
-export { default } from './confirm-approve.container'
+export { default } from './confirm-approve'
